refactored render config and fixed issue with 100% percentile marker
